trying to fix git case sensitive naming issues
